Don't navigate away from artefact on change.
Modifying the data triggers an update to the first navigation
selector.  This causes a "currentIndexChanged" signal to be emitted by
the first selector and received by the second.  If the slot accepts an
argument, the argument is the newly-selected index.  When
NavigatorComboBox.update() was given an optional argument, this then
made it start receiving a number in that argument which generally
screwed everything up.